Finish refactor on TrackList
