+++ dashboard/src/components/TrackList.tsx
@@ -1,61 +1,65 @@
-import { useAppDispatch, useAppSelector } from "helpers/hooks";
-import { Button, Card, Stack } from "react-bootstrap";
-import {
-  resetTrack,
-  untrackAlliance,
-  untrackGuild,
-  untrackPlayer,
-} from "store/track";
-import { Limits } from "types";
-import TrackListItem from "./TrackListItem";
+import { Badge, Button, ButtonGroup, ListGroup, Stack } from "react-bootstrap";
+import { TrackList as ITrackList } from "types";
 
-interface ITrackListProps {
-  limits: Limits;
-  onUpdateClick: React.MouseEventHandler<HTMLButtonElement>;
+interface ITrackListItemProps {
+  title: string;
+  limit: number;
+  list: ITrackList["players" | "guilds" | "alliances"];
+  onUntrackClick: (
+    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
+    id: string
+  ) => void;
 }
 
-const TrackList = ({ limits, onUpdateClick }: ITrackListProps) => {
-  const track = useAppSelector((state) => state.track);
-  const dispatch = useAppDispatch();
-
+const TrackList = ({
+  title,
+  limit = 0,
+  list,
+  onUntrackClick,
+}: ITrackListItemProps) => {
   return (
-    <Card>
-      <Stack gap={3} className="p-2">
-        <TrackListItem
-          title="Players"
-          limit={limits.players}
-          list={track.players}
-          onUntrackClick={(e, id) => dispatch(untrackPlayer(id))}
-        />
-        <TrackListItem
-          title="Guilds"
-          limit={limits.guilds}
-          list={track.guilds}
-          onUntrackClick={(e, id) => dispatch(untrackGuild(id))}
-        />
-        <TrackListItem
-          title="Alliances"
-          limit={limits.alliances}
-          list={track.alliances}
-          onUntrackClick={(e, id) => dispatch(untrackAlliance(id))}
-        />
-      </Stack>
+    <ListGroup>
+      <ListGroup.Item
+        variant="primary"
+        className="d-flex justify-content-between align-items-baseline"
+      >
+        <div>{title}</div>
+        <Badge bg={list.length < limit ? "secondary" : "danger"}>
+          {list.length}/{limit}
+        </Badge>
+      </ListGroup.Item>
 
-      <Card.Footer className="d-flex justify-content-end align-items-center">
-        <Button
-          variant="secondary"
-          onClick={() => {
-            dispatch(resetTrack());
-          }}
-        >
-          Reset
-        </Button>
-        <div className="px-2" />
-        <Button variant="primary" onClick={onUpdateClick}>
-          Save
-        </Button>
-      </Card.Footer>
-    </Card>
+      {list.map(({ id, name, server }, i) => (
+        <ListGroup.Item key={id} className="paper">
+          <div className="d-flex justify-content-between align-items-center">
+            <Stack>
+              <span className="id-text">#{id}</span>
+              <Stack
+                direction="horizontal"
+                gap={2}
+                className="d-flex align-items-baseline"
+              >
+                <div className={i >= limit ? "text-danger" : ""}>{name}</div>
+                <Badge bg={server.toLowerCase().replace(" ", "-")}>
+                  {server}
+                </Badge>
+              </Stack>
+            </Stack>
+            <ButtonGroup size="sm">
+              <Button
+                size="sm"
+                variant="danger"
+                onClick={(e) => {
+                  return onUntrackClick(e, id);
+                }}
+              >
+                Remove
+              </Button>
+            </ButtonGroup>
+          </div>
+        </ListGroup.Item>
+      ))}
+    </ListGroup>
   );
 };
