+++ dashboard/src/store/track.ts
@@ -1,13 +1,22 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { TrackList } from "types";
 
-const initialState: TrackList = {
+interface ITrackState extends TrackList {
+  default: TrackList;
+  changed: boolean;
+}
+
+const initialState: ITrackState = {
   players: [],
   guilds: [],
   alliances: [],
+  default: {
+    players: [],
+    guilds: [],
+    alliances: [],
+  },
+  changed: false,
 };
-
-let defaultState: TrackList = { ...initialState };
 
 export const trackslice = createSlice({
   name: "track",
@@ -18,14 +27,18 @@ export const trackslice = createSlice({
       state.guilds = payload.guilds;
       state.alliances = payload.alliances;
 
-      defaultState.players = payload.players;
-      defaultState.guilds = payload.guilds;
-      defaultState.alliances = payload.alliances;
+      state.default.players = payload.players;
+      state.default.guilds = payload.guilds;
+      state.default.alliances = payload.alliances;
+
+      state.changed = false;
     },
     resetTrack: (state) => {
-      state.players = defaultState.players;
-      state.guilds = defaultState.guilds;
-      state.alliances = defaultState.alliances;
+      state.players = state.default.players;
+      state.guilds = state.default.guilds;
+      state.alliances = state.default.alliances;
+
+      state.changed = false;
     },
     trackPlayer: (
       state,
@@ -33,6 +46,7 @@ export const trackslice = createSlice({
     ) => {
       if (state.players.find(({ id }) => id === payload.id)) return;
       state.players.push(payload);
+      state.changed = true;
     },
     trackGuild: (
       state,
@@ -40,6 +54,7 @@ export const trackslice = createSlice({
     ) => {
       if (state.guilds.find(({ id }) => id === payload.id)) return;
       state.guilds.push(payload);
+      state.changed = true;
     },
     trackAlliance: (
       state,
@@ -47,21 +62,25 @@ export const trackslice = createSlice({
     ) => {
       if (state.alliances.find(({ id }) => id === payload.id)) return;
       state.alliances.push(payload);
+      state.changed = true;
     },
     untrackPlayer: (state, { payload }: PayloadAction<string>) => {
       const i = state.players.findIndex(({ id }) => id === payload);
       if (i < 0) return;
       state.players.splice(i, 1);
+      state.changed = true;
     },
     untrackGuild: (state, { payload }: PayloadAction<string>) => {
       const i = state.guilds.findIndex(({ id }) => id === payload);
       if (i < 0) return;
       state.guilds.splice(i, 1);
+      state.changed = true;
     },
     untrackAlliance: (state, { payload }: PayloadAction<string>) => {
       const i = state.alliances.findIndex(({ id }) => id === payload);
       if (i < 0) return;
       state.alliances.splice(i, 1);
+      state.changed = true;
     },
   },
 });
